+++ playwright.config.ts
@@ -19,7 +19,7 @@ export default defineConfig({
   ],
   webServer: {
     command: 'pnpm run dev',
-    url: 'http://localhost:5173/test/index.html',
+    url: 'http://localhost:5173/test/browser/index.html',
     reuseExistingServer: !process.env.CI
   }
 })